Update DG, UG and refactor code - Add class diagram for Expenses - Update architecture diagram to remove class icons - Make FileDataParser to implement CommandParser to enhance coherence - Put AppServices into expense.service package
