+++ src/main/java/fintrek/FinTrek.java
@@ -2,6 +2,7 @@ package fintrek;
 
 import fintrek.expense.core.RecurringExpenseManager;
 import fintrek.expense.core.RegularExpenseManager;
+import fintrek.expense.service.AppServices;
 import fintrek.misc.MessageDisplayer;
 import fintrek.parser.CommandRouter;
 import fintrek.parser.RouteResult;
