+++ src/main/java/fintrek/parser/FileDataParser.java
@@ -7,20 +7,24 @@ import fintrek.util.InputValidator;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.util.regex.Pattern;
 
 /**
  * Parses saved expense file data into an Expense object.
  */
-public class FileDataParser {
+public class FileDataParser implements CommandParser<ParseResult<Void>> {
 
-    private static final Pattern AMOUNT_PATTERN = Pattern.compile("\\$(\\d+(?:\\.\\d{1,2})?)");
-    private static final String FORMAT_HINT = "Invalid expense format in save file.";
+    private static final FileDataParser INSTANCE = new FileDataParser(); // singleton
 
     public static ParseResult<Void> parseFileData(String fileData) {
+        return INSTANCE.parse(fileData); // delegates to the instance method
+    }
+
+    @Override
+    public ParseResult<Void> parse(String fileData) {
         if (InputValidator.isNullOrBlank(fileData)) {
             return ParseResult.failure(MessageDisplayer.EMPTY_DATA_MESSAGE);
         }
+
         String[] tokens = fileData.trim().split("\\|", 4);
         if (tokens.length < 2) {
             return ParseResult.failure(MessageDisplayer.EMPTY_AMOUNT_DATA_MESSAGE);
@@ -28,30 +32,30 @@ public class FileDataParser {
         if (tokens.length < 3) {
             return ParseResult.failure(MessageDisplayer.EMPTY_CATEGORY_DATA_MESSAGE);
         }
-        if(tokens.length < 4) {
+        if (tokens.length < 4) {
             return ParseResult.failure(MessageDisplayer.EMPTY_DATE_DATA_MESSAGE);
         }
 
         return processExpense(tokens);
     }
 
-    private static ParseResult<Void> processExpense(String[] tokens) {
+    private ParseResult<Void> processExpense(String[] tokens) {
         String description = tokens[0].trim();
-        String amountStr = (tokens[1].trim()).substring(1);
+        String amountStr = tokens[1].trim().substring(1);
         String category = tokens[2].trim();
         String dateStr = tokens[3].trim();
 
-        if(InputValidator.isNullOrBlank(description)) {
+        if (InputValidator.isNullOrBlank(description)) {
             return ParseResult.failure(MessageDisplayer.EMPTY_DESC_DATA_MESSAGE);
         }
 
-        if(!InputValidator.isValidPositiveDouble(amountStr)) {
+        if (!InputValidator.isValidPositiveDouble(amountStr)) {
             return ParseResult.failure(MessageDisplayer.INVALID_AMT_DATA_MESSAGE);
         }
 
         double amount = Double.parseDouble(amountStr);
 
-        if(!InputValidator.isValidDate(dateStr)) {
+        if (!InputValidator.isValidDate(dateStr)) {
             return ParseResult.failure(MessageDisplayer.INVALID_DATE_DATA_MESSAGE);
         }
 
@@ -60,5 +64,4 @@ public class FileDataParser {
         RegularExpenseManager.getInstance().add(newExpense);
         return ParseResult.success(null);
     }
-
 }
